Enhancing the carriage mod
- option "enable_brim" to add a lower/upper brim to the carriage. The
punch-out in the brim is smaller than the bearing diameter, so it should
be able to use it w/o zip-ties.
- support for bearing types LM10UU, RJ4JP, RJUM1. controled by variable "btype"
- changed actuator to a triangle base shape. Also changed the position of the
actuator - it's now mounted from top of the carriage. So no need to disassemble
the printer when playing with the max. endstop height.
- always build with "bonus" to get minimized size
- refactoring
TODO: modify the belt connector block - with current shape the belt sometimes
touches the block